End-of-day commit

+++ src/hashtab.c
@@ -126,7 +126,7 @@ HNODEPTR new_hnode(char *str)
    {
       if (verbose)
       {
-         fprintf(stderr,"[new_hnode] %s (%d)",msg_big_one,strlen(str));
+         fprintf(stderr,"[new_hnode] %s (%ld)",msg_big_one,(long)strlen(str));
          if (debug_mode)
             fprintf(stderr,":\n--> %s",str);
          fprintf(stderr,"\n");
@@ -316,7 +316,7 @@ UNODEPTR new_unode(char *str)
    {
       if (verbose)
       {
-         fprintf(stderr,"[new_unode] %s (%d)",msg_big_one,strlen(str));
+         fprintf(stderr,"[new_unode] %s (%ld)",msg_big_one,(long)strlen(str));
          if (debug_mode)
             fprintf(stderr,":\n--> %s",str);
          fprintf(stderr,"\n");
@@ -443,7 +443,7 @@ RNODEPTR new_rnode(char *str)
    {
       if (verbose)
       {
-         fprintf(stderr,"[new_rnode] %s (%d)",msg_big_one,strlen(str));
+         fprintf(stderr,"[new_rnode] %s (%ld)",msg_big_one,(long)strlen(str));
          if (debug_mode)
             fprintf(stderr,":\n--> %s",str);
          fprintf(stderr,"\n");
@@ -562,7 +562,7 @@ ANODEPTR new_anode(char *str)
    {
       if (verbose)
       {
-         fprintf(stderr,"[new_anode] %s (%d)",msg_big_one,strlen(str));
+         fprintf(stderr,"[new_anode] %s (%ld)",msg_big_one,(long)strlen(str));
          if (debug_mode)
             fprintf(stderr,":\n--> %s",str);
          fprintf(stderr,"\n");
@@ -678,7 +678,7 @@ SNODEPTR new_snode(char *str)
    {
       if (verbose)
       {
-         fprintf(stderr,"[new_snode] %s (%d)",msg_big_one,strlen(str));
+         fprintf(stderr,"[new_snode] %s (%ld)",msg_big_one,(long)strlen(str));
          if (debug_mode)
             fprintf(stderr,":\n--> %s",str);
          fprintf(stderr,"\n");
@@ -783,7 +783,7 @@ INODEPTR new_inode(char *str)
    {
       if (verbose)
       {
-         fprintf(stderr,"[new_inode] %s (%d)",msg_big_one,strlen(str));
+         fprintf(stderr,"[new_inode] %s (%ld)",msg_big_one,(long)strlen(str));
          if (debug_mode)
             fprintf(stderr,":\n--> %s",str);
          fprintf(stderr,"\n");
